requirements.txt
Flask
flask-cors
python-dotenv
openai
groq
faiss-cpu
sentence-transformers
pydub
requests
python-multipart
